lint: split rule visitor tests
